# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VidyaShala API specification - India's AI-powered multilingual learning platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /creators:
    get:
      operationId: listCreators
      tags: [creators]
      summary: List all creators
      parameters:
        - name: sector
          in: query
          schema:
            type: string
        - name: verified
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of creators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creator"
    post:
      operationId: createCreator
      tags: [creators]
      summary: Create a creator channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreatorBody"
      responses:
        "201":
          description: Created creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
  /creators/{id}:
    get:
      operationId: getCreator
      tags: [creators]
      summary: Get a creator's channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Creator channel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
  /creators/{id}/analytics:
    get:
      operationId: getCreatorAnalytics
      tags: [creators]
      summary: Get creator analytics dashboard data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorAnalytics"
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List quizzes
      parameters:
        - name: sector
          in: query
          schema:
            type: string
        - name: videoId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a new quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuizBody"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get a specific quiz with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}/submit:
    post:
      operationId: submitQuizAttempt
      tags: [quizzes]
      summary: Submit quiz answers and get score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmitBody"
      responses:
        "200":
          description: Quiz result with score and analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
  /community/posts:
    get:
      operationId: listCommunityPosts
      tags: [community]
      summary: List community discussion posts
      parameters:
        - name: sector
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Community posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityPost"
    post:
      operationId: createCommunityPost
      tags: [community]
      summary: Create a community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/posts/{id}/comments:
    get:
      operationId: getPostComments
      tags: [community]
      summary: Get comments for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addComment
      tags: [community]
      summary: Add a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /ai/ask:
    post:
      operationId: askAI
      tags: [ai]
      summary: Ask the AI tutor a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskAIBody"
      responses:
        "200":
          description: AI tutor response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIResponse"
  /ai/generate-quiz:
    post:
      operationId: generateAIQuiz
      tags: [ai]
      summary: Generate a quiz using AI for a given topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuizBody"
      responses:
        "200":
          description: Generated quiz questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedQuiz"
  /ai/flashcards:
    post:
      operationId: generateFlashcards
      tags: [ai]
      summary: Generate flashcards for a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardsBody"
      responses:
        "200":
          description: Generated flashcards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"
  /ai/study-plan:
    post:
      operationId: generateStudyPlan
      tags: [ai]
      summary: Generate a personalized study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanBody"
      responses:
        "200":
          description: Study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
...
components:
  schemas:
...
    Video:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        videoUrl:
          type: string
        duration:
          type: string
        views:
          type: integer
        likes:
          type: integer
        sector:
          type: string
        language:
          type: string
        tags:
          type: array
          items:
            type: string
        creatorId:
          type: integer
        creatorName:
          type: string
        creatorAvatar:
          type: string
        isVerified:
          type: boolean
        hasQuiz:
          type: boolean
        hasNotes:
          type: boolean
        hasFlashcards:
          type: boolean
        chapterCount:
          type: integer
        difficulty:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - thumbnailUrl
        - duration
        - views
        - likes
        - sector
        - language
        - creatorId
        - creatorName
        - creatorAvatar
        - isVerified
        - hasQuiz
        - hasNotes
        - hasFlashcards
        - chapterCount
        - difficulty
        - createdAt
...
    Creator:
      type: object
      properties:
        id:
          type: integer
        channelName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        sector:
          type: string
        isVerified:
          type: boolean
        followers:
          type: integer
        totalViews:
          type: integer
        videoCount:
          type: integer
        shortCount:
          type: integer
        languages:
          type: array
          items:
            type: string
        joinedAt:
          type: string
        monthlyEarnings:
          type: number
      required:
        - id
        - channelName
        - avatarUrl
        - sector
        - isVerified
        - followers
        - totalViews
        - videoCount
        - shortCount
        - languages
        - joinedAt
        - monthlyEarnings
    CreateCreatorBody:
      type: object
      properties:
        channelName:
          type: string
        bio:
          type: string
        sector:
          type: string
        languages:
          type: array
          items:
            type: string
      required:
        - channelName
        - sector
        - languages
    CreatorAnalytics:
      type: object
      properties:
        creatorId:
          type: integer
        totalViews:
          type: integer
        totalLikes:
          type: integer
        totalSubscribers:
          type: integer
        monthlyEarnings:
          type: number
        viewsByMonth:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyData"
        earningsByMonth:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyData"
        topVideos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        engagementRate:
          type: number
      required:
        - creatorId
        - totalViews
        - totalLikes
        - totalSubscribers
        - monthlyEarnings
        - viewsByMonth
        - earningsByMonth
        - topVideos
        - engagementRate
    MonthlyData:
      type: object
      properties:
        month:
          type: string
        value:
          type: number
      required:
        - month
        - value
...
    Quiz:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        sector:
          type: string
        difficulty:
          type: string
        questionCount:
          type: integer
        timeLimit:
          type: integer
        passingScore:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        videoId:
          type: ["integer", "null"]
        creatorId:
          type: integer
        attempts:
          type: integer
        averageScore:
          type: number
      required:
        - id
        - title
        - sector
        - difficulty
        - questionCount
        - timeLimit
        - passingScore
        - questions
        - creatorId
        - attempts
        - averageScore
    QuizQuestion:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        explanation:
          type: string
      required:
        - id
        - question
        - options
        - correctIndex
        - explanation
    CreateQuizBody:
      type: object
      properties:
        title:
          type: string
        sector:
          type: string
        difficulty:
          type: string
        timeLimit:
          type: integer
        passingScore:
          type: integer
        videoId:
          type: ["integer", "null"]
        creatorId:
          type: integer
      required:
        - title
        - sector
        - difficulty
        - timeLimit
        - passingScore
        - creatorId
    QuizSubmitBody:
      type: object
      properties:
        answers:
          type: array
          items:
            type: integer
        timeTaken:
          type: integer
      required:
        - answers
        - timeTaken
    QuizResult:
      type: object
      properties:
        quizId:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        passed:
          type: boolean
        timeTaken:
          type: integer
        badge:
          type: ["string", "null"]
        weakTopics:
          type: array
          items:
            type: string
        nextSuggestions:
          type: array
          items:
            type: string
      required:
        - quizId
        - score
        - totalQuestions
        - correctAnswers
        - passed
        - timeTaken
        - weakTopics
        - nextSuggestions
...
    CommunityPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        sector:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: string
        likes:
          type: integer
        commentCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - content
        - sector
        - authorName
        - authorAvatar
        - likes
        - commentCount
        - tags
        - createdAt
    CreatePostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        sector:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - content
        - sector
    Comment:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: string
        likes:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - content
        - authorName
        - authorAvatar
        - likes
        - createdAt
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    AskAIBody:
      type: object
      properties:
        question:
          type: string
        language:
          type: string
        context:
          type: string
        mode:
          type: string
      required:
        - question
        - language
    AIResponse:
      type: object
      properties:
        answer:
          type: string
        language:
          type: string
        suggestions:
          type: array
          items:
            type: string
        relatedTopics:
          type: array
          items:
            type: string
        mode:
          type: string
      required:
        - answer
        - language
        - suggestions
        - relatedTopics
        - mode
    GenerateQuizBody:
      type: object
      properties:
        topic:
          type: string
        sector:
          type: string
        difficulty:
          type: string
        questionCount:
          type: integer
        language:
          type: string
      required:
        - topic
        - sector
        - difficulty
        - questionCount
        - language
    GeneratedQuiz:
      type: object
      properties:
        topic:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
      required:
        - topic
        - questions
    FlashcardsBody:
      type: object
      properties:
        topic:
          type: string
        language:
          type: string
        count:
          type: integer
      required:
        - topic
        - language
        - count
    Flashcard:
      type: object
      properties:
        front:
          type: string
        back:
          type: string
        topic:
          type: string
      required:
        - front
        - back
        - topic
    StudyPlanBody:
      type: object
      properties:
        goal:
          type: string
        sector:
          type: string
        hoursPerDay:
          type: integer
        durationWeeks:
          type: integer
        language:
          type: string
      required:
        - goal
        - sector
        - hoursPerDay
        - durationWeeks
        - language
    StudyPlan:
      type: object
      properties:
        goal:
          type: string
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/StudyWeek"
        totalHours:
          type: integer
        milestones:
          type: array
          items:
            type: string
      required:
        - goal
        - weeks
        - totalHours
        - milestones
    StudyWeek:
      type: object
      properties:
        week:
          type: integer
        theme:
          type: string
        topics:
          type: array
          items:
            type: string
        hoursRequired:
          type: integer
      required:
        - week
        - theme
        - topics
        - hoursRequired
...