# lib/api-zod/src/generated/types/flashcard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VidyaShala API specification - India's AI-powered multilingual learning platform
 * OpenAPI spec version: 0.1.0
 */

export interface Flashcard {
  front: string;
  back: string;
  topic: string;
}