# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VidyaShala API specification - India's AI-powered multilingual learning platform
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  title: string;
  description?: string;
  thumbnailUrl: string;
  videoUrl?: string;
  duration: string;
  views: number;
  likes: number;
  sector: string;
  language: string;
  tags?: string[];
  creatorId: number;
  creatorName: string;
  creatorAvatar: string;
  isVerified: boolean;
  hasQuiz: boolean;
  hasNotes: boolean;
  hasFlashcards: boolean;
  chapterCount: number;
  difficulty: string;
  createdAt: string;
}